tiny

+++ src/context/entity.h
@@ -6,8 +6,8 @@ namespace context {
 
 struct Brick {
 	int id = 0;
-	int x = 0;
-	int y = 0;
+	float x = 0;
+	float y = 0;
 	int region = 0;
 };
